+++ src/App.tsx
@@ -2,7 +2,7 @@ import rcoiLogo from "./assets/rcoi.jpg";
 
 import Colors from "./components/Colors.js";
 import Counter from "./components/Counter.js";
-import IsInput from "./components/IsInput.tsx";
+
 import Task1 from "./components/Task1.tsx";
 import Task2 from "./components/Task2.tsx";
 import Task3 from "./components/Task3.tsx";
@@ -10,7 +10,6 @@ import Task4 from "./components/Task4.tsx";
 import Task5 from "./components/Task5.tsx";
 import Task6 from "./components/Task6.tsx";
 import Task7 from "./components/Task7.tsx";
-import Task8 from "./components/Task8.tsx";
 
 import "./App.css";
 
@@ -19,7 +18,6 @@ function App() {
     <div>
       <Colors />
       <Counter />
-      <IsInput />
       <Task1 />
       <Task2 />
       <Task3 />
@@ -27,8 +25,6 @@ function App() {
       <Task5 />
       <Task6 />
       <Task7 />
-      <Task8 />
-
       <h1 className="title">
         ГКУ "Региональный центр обработки информации ЕГЭ и мониторинга качества
         образования Чеченской Республики"
