+++ src/components/Task5.tsx
@@ -1,23 +1,40 @@
-// 2. Создай компонент с кнопкой и заголовком (например, <h1>). При каждом клике на кнопку текст в заголовке должен изменяться на случайную фразу из массива (например, массив из пяти разных фраз).
+// 5. Создай компонент с текстовым полем (input) и списком имен (например, массив из 5-10 имен). При вводе текста в поле список должен фильтроваться по именам, которые содержат введенные буквы.
 
 import { useState } from "react";
 
-const phrases = [
-  "Хавашка учится с душой",
-  "Код Хавашки — чистый и аккуратный",
-  "Хавашка не сдаётся, даже если сложно",
-  "Каждый день Хавашка становится сильнее",
-  "Frontend — это путь, который Хавашка точно пройдёт",
+const names = [
+  "Хавашка",
+  "Эмма",
+  "Оливия",
+  "Ава",
+  "София",
+  "Изабелла",
+  "Миа",
+  "Шарлотта",
+  "Амелия",
+  "Харпер",
 ];
 const Task5 = () => {
-  const [toggle, setToggle] = useState(0);
-  const handleToggle = () => {
-    setToggle((toggle + 1) % phrases.length);
-  };
+  const [search, setSearch] = useState("");
+
+  const filtered = names.filter((name) =>
+    name.toLowerCase().includes(search.toLowerCase())
+  );
   return (
     <div>
-      <p>{phrases[toggle]}</p>
-      <button onClick={handleToggle}> Изменить фразу</button>
+      <input
+        type="text"
+        value={search}
+        onChange={(e) => setSearch(e.target.value)}
+        placeholder="Поиск"
+      />
+      <div>
+        {filtered.length > 0 ? (
+          filtered.map((name) => <p>{name}</p>)
+        ) : (
+          <p>Ничего не найдено</p>
+        )}
+      </div>
     </div>
   );
 };
